requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
fastapi==0.95.0
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
pillow==9.5.0
pydantic==1.10.7
python-multipart==0.0.6
requests==2.28.2
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.26.1 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn==0.21.1
